requirements.txt
crawl4ai
asyncio
playwright
fastapi[standard]
pydantic
PyJWT
python-dotenv